+++ src/main/java/com/archyx/aureliumskills/mana/MAbility.java
@@ -26,7 +26,8 @@ public enum MAbility implements AbstractAbility {
     SHARP_HOOK(() -> Skills.FISHING, 0.5, 0.5, 2, -0.1, 5, 2,
             new String[] {"display_damage_with_scaling", "enable_sound"}, new Object[] {true, true}),
     TERRAFORM(() -> Skills.EXCAVATION, 5.0, 4.0, 200, -5, 20, 20,
-            new String[] {"require_sneak", "check_offhand", "sneak_offhand_bypass"}, new Object[] {false, true, true}),
+            new String[] {"require_sneak", "check_offhand", "sneak_offhand_bypass"}, new Object[] {false, true, true},
+            Terraform.class),
     CHARGED_SHOT(() -> Skills.ARCHERY, 0.5, 0.3, 0, 0, 5, 5,
             new String[] {"enable_message", "enable_sound"}, new Object[] {true, true},
             ChargedShot.class),
