+++ src/main/java/com/archyx/aureliumskills/skills/excavation/ExcavationAbilities.java
@@ -5,37 +5,19 @@ import com.archyx.aureliumskills.ability.Ability;
 import com.archyx.aureliumskills.ability.AbilityProvider;
 import com.archyx.aureliumskills.api.event.LootDropCause;
 import com.archyx.aureliumskills.api.event.PlayerLootDropEvent;
-import com.archyx.aureliumskills.api.event.TerraformBlockBreakEvent;
 import com.archyx.aureliumskills.configuration.OptionL;
 import com.archyx.aureliumskills.data.PlayerData;
-import com.archyx.aureliumskills.lang.Lang;
-import com.archyx.aureliumskills.lang.ManaAbilityMessage;
-import com.archyx.aureliumskills.mana.MAbility;
-import com.archyx.aureliumskills.mana.ManaAbilityManager;
-import com.archyx.aureliumskills.mana.Terraform;
 import com.archyx.aureliumskills.skills.Skills;
-import com.archyx.aureliumskills.source.SourceTag;
-import com.archyx.aureliumskills.util.math.NumberUtil;
-import com.archyx.aureliumskills.util.text.TextUtil;
 import org.bukkit.Bukkit;
 import org.bukkit.GameMode;
 import org.bukkit.Material;
 import org.bukkit.block.Block;
-import org.bukkit.block.BlockFace;
 import org.bukkit.enchantments.Enchantment;
 import org.bukkit.entity.Player;
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
-import org.bukkit.event.block.Action;
-import org.bukkit.event.block.BlockBreakEvent;
 import org.bukkit.event.entity.EntityDamageByEntityEvent;
-import org.bukkit.event.player.PlayerInteractEvent;
 import org.bukkit.inventory.ItemStack;
-import org.bukkit.metadata.FixedMetadataValue;
 
-import java.util.LinkedList;
-import java.util.Locale;
 import java.util.Random;
 
 public class ExcavationAbilities extends AbilityProvider implements Listener {
@@ -96,93 +78,4 @@ public class ExcavationAbilities extends AbilityProvider implements Listener {
 			}
 		}
 	}
-
-	@EventHandler(priority = EventPriority.HIGH)
-	public void excavationListener(BlockBreakEvent event) {
-		if (OptionL.isEnabled(Skills.EXCAVATION)) {
-			if (!event.isCancelled()) {
-				Player player = event.getPlayer();
-				Block block = event.getBlock();
-				if (blockAbility(player)) return;
-				//Applies abilities
-				PlayerData playerData = plugin.getPlayerManager().getPlayerData(player);
-				if (playerData == null) return;
-				if (plugin.getAbilityManager().isEnabled(MAbility.TERRAFORM)) {
-					if (!block.hasMetadata("AureliumSkills-Terraform") && event.getClass() == BlockBreakEvent.class) {
-						applyTerraform(player, block);
-					}
-				}
-			}
-		}
-	}
-
-	private void applyTerraform(Player player, Block block) {
-		ManaAbilityManager manager = plugin.getManaAbilityManager();
-		// Check if block is applicable to ability
-		ExcavationSource source = ExcavationSource.getSource(block);
-		if (source == null) return;
-		if (!hasTag(source, SourceTag.TERRAFORM_APPLICABLE)) return;
-		// Apply if activated
-		if (manager.isActivated(player.getUniqueId(), MAbility.TERRAFORM)) {
-			terraformBreak(player, block);
-			return;
-		}
-		//Checks if speed mine is ready
-		if (manager.isReady(player.getUniqueId(), MAbility.TERRAFORM)) {
-			//Checks if holding pickaxe
-			Material mat = player.getInventory().getItemInMainHand().getType();
-			if (mat.name().contains("SHOVEL") || mat.name().contains("SPADE")) {
-				PlayerData playerData = plugin.getPlayerManager().getPlayerData(player);
-				if (playerData == null) return;
-				Locale locale = playerData.getLocale();
-				if (playerData.getMana() >= getManaCost(MAbility.TERRAFORM, playerData)) {
-					manager.activateAbility(player, MAbility.TERRAFORM, (int) (getValue(MAbility.TERRAFORM, playerData) * 20), new Terraform(plugin));
-					terraformBreak(player, block);
-				}
-				else {
-					plugin.getAbilityManager().sendMessage(player, TextUtil.replace(Lang.getMessage(ManaAbilityMessage.NOT_ENOUGH_MANA, locale)
-							,"{mana}", NumberUtil.format0(manager.getManaCost(MAbility.TERRAFORM, playerData))
-							, "{current_mana}", String.valueOf(Math.round(playerData.getMana()))
-							, "{max_mana}", String.valueOf(Math.round(playerData.getMaxMana()))));
-				}
-			}
-		}
-	}
-
-	@EventHandler
-	private void readyTerraform(PlayerInteractEvent event) {
-		plugin.getManaAbilityManager().getActivator().readyAbility(event, Skills.EXCAVATION, new String[] {"SHOVEL", "SPADE"}, Action.RIGHT_CLICK_AIR, Action.RIGHT_CLICK_BLOCK);
-	}
-
-	private void terraformBreak(Player player, Block block) {
-		Material material = block.getType();
-		BlockFace[] faces = new BlockFace[] {BlockFace.NORTH, BlockFace.SOUTH, BlockFace.EAST, BlockFace.WEST};
-		LinkedList<Block> toCheck = new LinkedList<>();
-		toCheck.add(block);
-		int count = 0;
-		while ((block = toCheck.poll()) != null && count < 61) {
-			if (block.getType() == material) {
-				block.setMetadata("AureliumSkills-Terraform", new FixedMetadataValue(plugin, true));
-				breakBlock(player, block);
-				for (BlockFace face : faces) {
-					toCheck.add(block.getRelative(face));
-				}
-				count++;
-			}
-		}
-	}
-
-	private void breakBlock(Player player, Block block) {
-		if (!plugin.getTownySupport().canBreak(player, block)) {
-			block.removeMetadata("AureliumSkills-Terraform", plugin);
-			return;
-		}
-		TerraformBlockBreakEvent event = new TerraformBlockBreakEvent(block, player);
-		Bukkit.getPluginManager().callEvent(event);
-		if (!event.isCancelled()) {
-			block.breakNaturally(player.getInventory().getItemInMainHand());
-		}
-		block.removeMetadata("AureliumSkills-Terraform", plugin);
-	}
-
 }
